rubah, fix reconfigure, fix postinstall

+++ src/module-utils/reconfigure.ts
@@ -1,7 +1,8 @@
 import fs from 'fs'
 import path from 'path'
 import readline from 'readline'
-import {buildconfig} from './tsconfig'
+import {buildconfig} from './templates/tsconfig'
+import cp from 'child_process'
 
 const crl = () => readline.createInterface({
     input: process.stdin,
@@ -15,13 +16,6 @@ async function reconfigure() {
     
     //SETTING UP RECONFIGURE SCRIPT
     let packageJson: any = JSON.parse(fs.readFileSync(path.join(cwd,"package.json")).toString());
-    packageJson.scripts.reconfigure = Object.keys(packageJson.averModule)
-        .sort((a,b)=>packageJson.averModule[a].priority - packageJson.averModule[a].priority)
-        .filter(x=>packageJson.averModule[x].reconfigure)
-        .filter(x=>x!=packageJson.name)
-        .map(x=>packageJson.averModule[x].reconfigure)
-        .concat(["echo finished reconfiguring"])
-        .join(" && ");
 
     if(packageJson.averModule.tsemplate.firstrun){
         delete packageJson.averModule.tsemplate.firstrun;
@@ -54,6 +48,7 @@ to install these dependencies and configure your project before continuing
     if(mode == "module") {
         if(!fs.existsSync(path.join(cwd,"tsconfig.build.json")))
             fs.writeFileSync(path.join(cwd,"tsconfig.build.json"),JSON.stringify(buildconfig,null,2));
+        packageJson.scripts.reconfigure = "node -r ts-node/register src/module-utils/reconfigure.ts";
         packageJson.scripts['build:ts.d'] = "tsc -d --project tsconfig.build.json --emitDeclarationOnly";
         packageJson.scripts['build:ts'] = "tsc --project tsconfig.build.json";
         packageJson.scripts.build = "rm -rf build && mkdir build && npm run build:ts && npm run build:ts.d";
@@ -61,18 +56,22 @@ to install these dependencies and configure your project before continuing
         packageJson.main = "build/index.js";
         packageJson.bin = "build/index.js";
         packageJson.types = "build/index.d.js";
-        let index = fs.readFileSync(path.join(cwd,"src","index.ts")).toString();
-        let i = index.indexOf("//your command line parsing logic here")+38;
-        index = index.substr(0,i)+"\n        case 'reconfigure': require('./module-utils/reconfigure'); break;"
-            +index.substr(i);
-        fs.writeFileSync(path.join(cwd,"src","index.ts"),index);
         if(!fs.existsSync(path.join(cwd, "src", "module-utils"))) fs.mkdirSync(path.join(cwd, "src", "module-utils"));
-        fs.writeFileSync(path.join(cwd,"src","module-utils","postinstall.ts"),"//YOUR POSTINSTALL SCRIPT HERE");
-        fs.writeFileSync(path.join(cwd,"src","module-utils","reconfigure.ts"),"//YOUR RECONFIGURE SCRIPT HERE");
+        if(!fs.existsSync(path.join(cwd,"src","module-utils","postinstall.ts")))
+            fs.writeFileSync(path.join(cwd,"src","module-utils","postinstall.ts"),"//YOUR POSTINSTALL SCRIPT HERE");
+        if(!fs.existsSync(path.join(cwd,"src","module-utils","reconfigure.ts")))
+            fs.writeFileSync(path.join(cwd,"src","module-utils","reconfigure.ts"),"//YOUR RECONFIGURE SCRIPT HERE");
         packageJson.scripts.postinstall = "node -r ts-node/register src/module-utils/postinstall.ts && "+
             "node -r ts-node/register src/module-utils/reconfigure.ts";
     } else {
         // APPLICATION RECONFIGURATION GOES HERE
+    }
+
+    for(let moduleName in packageJson.averModule){
+        if(moduleName == packageJson.name) continue;
+        let pkgjson = JSON.parse(fs.readFileSync(`node_modules/${moduleName}/package.json`).toString());
+        console.log(`reconfiguring ${moduleName}`);
+        console.log(cp.execSync(`node_modules/${moduleName}/${pkgjson.main} reconfigure`).toString());
     }
 
     fs.writeFileSync(path.join(cwd,"package.json"),JSON.stringify(packageJson,null,2));
